refacto environment package
