+++ src/super_gradients/common/environment/omegaconf_utils.py
@@ -1,0 +1,38 @@
+import importlib
+import os
+import sys
+
+from omegaconf import OmegaConf
+
+from super_gradients.common.environment import environment_config
+
+
+def get_cls(cls_path: str):
+    """
+    A resolver for Hydra/OmegaConf to allow getting a class instead on an instance.
+    usage:
+    class_of_optimizer: ${class:torch.optim.Adam}
+    """
+    module = ".".join(cls_path.split(".")[:-1])
+    name = cls_path.split(".")[-1]
+    importlib.import_module(module)
+    return getattr(sys.modules[module], name)
+
+
+def hydra_output_dir_resolver(ckpt_root_dir: str, experiment_name: str) -> str:
+    if ckpt_root_dir is None:
+        output_dir_path = environment_config.PKG_CHECKPOINTS_DIR + os.path.sep + experiment_name
+    else:
+        output_dir_path = ckpt_root_dir + os.path.sep + experiment_name
+    return output_dir_path
+
+
+def register_hydra_resolvers():
+    """Register all the hydra resolvers required for the super-gradients recipes."""
+    OmegaConf.register_new_resolver("hydra_output_dir", hydra_output_dir_resolver, replace=True)
+    OmegaConf.register_new_resolver("class", lambda *args: get_cls(*args), replace=True)
+    OmegaConf.register_new_resolver("add", lambda *args: sum(args), replace=True)
+    OmegaConf.register_new_resolver("cond", lambda boolean, x, y: x if boolean else y, replace=True)
+    OmegaConf.register_new_resolver("getitem", lambda container, key: container[key], replace=True)  # get item from a container (list, dict...)
+    OmegaConf.register_new_resolver("first", lambda lst: lst[0], replace=True)  # get the first item from a list
+    OmegaConf.register_new_resolver("last", lambda lst: lst[-1], replace=True)  # get the last item from a list
